+++ score-junit/src/main/java/com/dojocoders/score/junit/persistence/TestPersistUnit.java
@@ -2,6 +2,6 @@ package com.dojocoders.score.junit.persistence;
 
 public interface TestPersistUnit {
 
-    void putScore(String team, Integer points);
+    void putScore(String team, Long points);
 
 }
