+++ score-junit/src/main/java/com/dojocoders/score/junit/persistence/TestStaticMap.java
@@ -6,10 +6,10 @@ import java.util.Map;
 
 public class TestStaticMap implements TestPersistUnit {
 
-    private static final Map<String, Integer> SCORE_MAP = Maps.newConcurrentMap();
+    private static final Map<String, Long> SCORE_MAP = Maps.newConcurrentMap();
 
     @Override
-    public void putScore(String team, Integer result) {
+    public void putScore(String team, Long result) {
         SCORE_MAP.put(team, result);
     }
 
@@ -17,7 +17,7 @@ public class TestStaticMap implements TestPersistUnit {
         SCORE_MAP.clear();
     }
 
-    public Map<String, Integer> getAllScores() {
+    public Map<String, Long> getAllScores() {
         return SCORE_MAP;
     }
 }
